move all files to submarine package
