--- TP3/submarine/Main.java
@@ -1,9 +1,0 @@
-package submarine;
-
-public class Main {
-    public static void main(String[] args) {
-        Submarine submarine = new Submarine();
-        submarine.executeInstructions("");
-        System.out.println(submarine.getAllCoordsAndOrientation());
-    }
-}
